// dotnet/openapi/model/courseModelServiceResponse.ts
/**
 * JobPortal Admin Production
 *
 * Contact: 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CourseModel } from './courseModel';


export interface CourseModelServiceResponse { 
    message?: string | null;
    statusCode?: number;
    data?: CourseModel;
}