// dotnet/openapi/model/createCourseCommand.ts
/**
 * JobPortal Admin Production
 *
 * Contact: 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateCourseCommand { 
    title?: string | null;
    description?: string | null;
    durationInWeeks?: number;
}