// dotnet/openapi/model/updateCourseCommand.ts
/**
 * JobPortal Admin Production
 *
 * Contact: 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateCourseCommand { 
    title?: string | null;
    description?: string | null;
    durationInWeeks?: number;
    id?: string;
}